+++ libr/include/r/util.h
@@ -32,7 +32,7 @@ extern "C" {
 #endif
 
 #define rstreq(_s1, _s2) ((_s1) != NULL && (_s2) != NULL && strcmp(_s1,_s2)==0)
-char *rprintf(char *buffer, int len, const char *format, ...);
+char *rprintf(char *buffer, size_t len, const char *format, ...);
 int r_random(void *buf, size_t buflen);
 char *r_uuid();
